requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
bottle==0.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
executing==2.2.0
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.10
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.11
lxml==6.0.0
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.99.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
proxy_tools==0.1.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyobjc-framework-Security==11.1
pyobjc-framework-WebKit==11.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pywebview==5.4
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
stack-data==0.6.3
streamlit==1.47.1
tenacity==9.1.2
tiktoken==0.10.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
zstandard==0.23.0
